move all gui files into gui folder
