+++ src/main/java/penguin/gui/Main.java
@@ -1,4 +1,4 @@
-package penguin;
+package penguin.gui;
 
 import java.io.IOException;
 
@@ -7,6 +7,7 @@ import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
 import javafx.scene.layout.AnchorPane;
 import javafx.stage.Stage;
+import penguin.app.Penguin;
 
 /**
  * A GUI for Duke using FXML.
