+++ src/main/java/penguin/gui/MainWindow.java
@@ -1,4 +1,4 @@
-package penguin;
+package penguin.gui;
 
 import javafx.fxml.FXML;
 
@@ -10,6 +10,7 @@ import javafx.scene.layout.AnchorPane;
 import javafx.scene.layout.VBox;
 
 import javafx.stage.Stage;
+import penguin.app.Penguin;
 
 /**
  * Controller for the main GUI.
